pyproject.toml
[project]
name = "airoleplay"
version = "0.1.0"
description = "AI-powered roleplay character system using deepagents"
requires-python = ">=3.10"
dependencies = [
    "deepagents>=0.1.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",  # For Whisper transcription
    "streamlit>=1.28.0",  # Web UI
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
